% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\docType{data}
\name{bmj}
\alias{bmj}
\title{Palette inspired by the British Medical Journal.}
\format{
An object of class \code{character} of length 9.
}
\usage{
bmj
}
\description{
Palette inspired by the British Medical Journal.
}
\keyword{datasets}
