% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_functions.R
\name{chisq.fisher}
\alias{chisq.fisher}
\title{Internal test for chi-squared assumption.Fisher (2 by 2). If results = T, it fails}
\usage{
chisq.fisher(tab)
}
\arguments{
\item{tab}{A numeric two by two table.}
}
\description{
\code{chisq.fisher} is an internal function called by \code{contingency} and \code{contingency2} that uses the Fisher exact test if results from the assumptions for the chi-squared test fail.
}
