% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_functions.R
\name{diag_test2}
\alias{diag_test2}
\title{Diagnostic tests from direct input.}
\usage{
diag_test2(aa, bb, cc, dd)
}
\arguments{
\item{aa}{Number of cases where both screening test and the gold standard are positive.}

\item{bb}{Number of cases where screening test is positive but gold standard is negative.}

\item{cc}{Number of cases where screening test is negative but gold standard is positive.}

\item{dd}{Number of cases where both screening test and the gold standard are negative.}
}
\description{
\code{diag_test2} is a wrap that calls \link{epi.tests} from package \code{epiR}.
It computes sensitivity, specificity and other statistics related with screening tests.
}
\details{
\code{diag.test} uses direct input variables.
}
\examples{
## We compare the use of lung’s X-rays on the screening of TB against the gold standard test.
diag_test2(22, 51, 8, 1739)
}
\seealso{
\code{epiR::epi.tests}
}
