% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{gf_star}
\alias{gf_star}
\title{Annotating a plot to display differences between groups.}
\usage{
gf_star(fig, x1, y1, x2, y2, y3, legend = "*", ...)
}
\arguments{
\item{fig}{A gformula object.}

\item{x1}{Position in \code{x} for the start of the horizontal line.}

\item{y1}{Position in \code{y} for the start of the vertical line, below to the horizontal line.}

\item{x2}{Position in \code{x} for the end of the horizontal line.}

\item{y2}{Position in \code{y} where the horizontal line is drawn.}

\item{y3}{Position in \code{y} where the text is added.}

\item{legend}{Character text used for annotating the plot.}

\item{...}{Additional information passed to \code{\link{gf_text}}.}
}
\description{
\code{gf_star} Is a function used to display differences between groups (see details).
}
\details{
This function draws an horizontal line from coordinate (\code{x1}, \code{y2})
to coordinate (\code{x2}, \code{y2}). Draws vertical lines below the horizontal line,
towards data, from (\code{x1}, \code{y1}) to (\code{x1}, \code{y2}) and from
(\code{x2}, \code{y1}) to (\code{x2}, \code{y2}). Finally, adds
text above the horizontal line, at the mid point between \code{x1} and \code{x2}.
See examples.
}
\examples{
data(kfm, package = "ISwR")
require(sjlabelled)
kfm = kfm \%>\%
  var_labels(
     dl.milk = 'Breast-milk intake (dl/day)',
     sex = 'Sex',
     weight = 'Child weight (kg)',
     ml.suppl = 'Milk substitute (ml/day)',
     mat.weight = 'Maternal weight (kg)',
     mat.height = 'Maternal height (cm)'
     )

kfm \%>\%
  box_plot(dl.milk ~ sex, fill='thistle', alpha = 0.8) \%>\%
  axis_labs() \%>\%
  gf_star(x1 = 1, y1 = 10.9, x2 = 2, y2 = 11, y3 = 11.2)

kfm \%>\%
  box_plot(dl.milk ~ sex, fill='thistle', alpha = 0.8) \%>\%
  axis_labs() \%>\%
  gf_star(1, 10.9, 2, 11, 11.4, legend = 'p = 0.035', size = 2.5)

data(energy, package = "ISwR")
energy = energy \%>\%
  var_labels(
    expend = 'Energy expenditure (MJ/day)',
    stature = 'Stature'
    )

energy \%>\%
  strip_error(expend ~ stature, col = 'red') \%>\%
  axis_labs() \%>\%
  gf_star(1, 13, 2, 13.2, 13.4, "**")
}
