% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_functions.R
\name{glmcoef}
\alias{glmcoef}
\title{Table of coefficients from generalised linear models (extended).}
\usage{
glmcoef(model, digits = 2, alpha = 0.05, labels = NULL, se.rob = TRUE)
}
\arguments{
\item{model}{A model from any of the classes listed in the details section.}

\item{digits}{A scalar, number of digits for rounding the results (default = 2).}

\item{alpha}{Significant level (default = 0.05) used to calculate confidence intervals.}

\item{labels}{An optional character vector with the names of the coefficients (including intercept).}

\item{se.rob}{Logical, should robust errors be used to calculate confidence intervals? (default = TRUE).}
}
\value{
A data frame with estimates, standard errors, confidence intervals and p-values from \code{glm} objects.
}
\description{
\code{glmcoef} is an internal function that displays estimates with confidence intervals and p-values from generalised linear models (see Details).
}
\details{
\code{glmcoef} recognises objects (models) from the following classes: \code{clm}, \code{clogit},
\code{coxph}, \code{gee}, \code{glm}, \code{glmerMod}, \code{lm}, \code{lme}, \code{multinom}, \code{negbin},
\code{polr} and \code{surveg}

For models from logistic regression (including conditional logistic, ordinal and multinomial),
Poisson or survival analysis, coefficient estimates and corresponding confidence intervals are
automatically exponentiated (back-transformed).

By default, \code{glmcoef} uses robust standard errors for calculating confidence intervals.

\code{glmcoef} is called by \code{coef_plot} and displays more information from regression
models than \code{glm_coef} which produces tables ready for publication.
}
\examples{
data(birthwt, package = "MASS")
birthwt$smoke <- factor(birthwt$smoke, labels=c("Non-smoker", "Smoker"))
birthwt$race <- factor(birthwt$race > 1, labels=c("White", "Non-white"))
model_norm <- glm(bwt ~ smoke + race, data = birthwt)
glmcoef(model_norm)
glmcoef(model_norm, labels=c("Constant", "Smoker vs Non-smoker", "Non-white vs White"))
}
\author{
Josie Athens, Department of Preventive and Social Medicine, University of Otago, New Zealand.
}
