% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{harm_mean}
\alias{harm_mean}
\title{Harmonic mean.}
\usage{
harm_mean(x)
}
\arguments{
\item{x}{A numeric variable with no zero values.}
}
\value{
A scalar, the calculated harmonic mean.
}
\description{
Harmonic mean.
}
\examples{
data(IgM, package = "ISwR")
Ab <- data.frame(IgM)
estat(~ IgM, data = Ab)
harm_mean(IgM)
}
\author{
Josie Athens, Department of Preventive and Social Medicine, University of Otago, New Zealand.
}
