% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse of the logit}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{Numerical value used to compute the inverse of the logit.}
}
\description{
\code{inv_logit} Calculates the inverse of the logit (probability in logistic regression)
}
\author{
Josie Athens, Department of Preventive and Social Medicine, University of Otago, New Zealand.
}
