% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{knife_mean}
\alias{knife_mean}
\title{Jackknife for means.}
\usage{
knife_mean(x)
}
\arguments{
\item{x}{A numerical variable. Missing values are removed for the mean calculation.}
}
\value{
A vector with the mean calculations.
}
\description{
\code{knife_mean} is an internal function. Calculates arithmetic means by removing one observation
at a time.
}
\examples{
x <- rnorm(10, 170, 8)
x
mean(x)
knife_mean(x)
}
