% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{leverage}
\alias{leverage}
\title{Leverage.}
\usage{
leverage(x)
}
\arguments{
\item{x}{A numeric variable. Missing values are removed by default.}
}
\value{
Variable with corresponding leverage estimations
}
\description{
\code{leverage} is an internal function called by \link{rank_leverage}.
}
\details{
Estimates the leverage of each observation around the arithmetic mean.
}
\examples{
x <- rnorm(10, 170, 8)
x
mean(x)
leverage(x)
rank_leverage(x)
}
