% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{panel.errbars}
\alias{panel.errbars}
\title{Internal function for displaying error bars in lattice plots.}
\usage{
panel.errbars(x, y, ..., panel.xy = panel.xyplot,
  make.grid = c("horizontal", "vertical", "both", "none"), ewidth = 0)
}
\arguments{
\item{x}{A numeric vector with the x positions for the bars.}

\item{y}{A numeric matrix with the mid values, lower values and upper values of the bars.}

\item{...}{Further arguments passed to lattice.}

\item{panel.xy}{A lattice panel.}

\item{make.grid}{Type of grid, a character with options: "horizontal", "vertical", "both" and "none".}

\item{ewidth}{An integer.}
}
\description{
\code{panel.errbars} is an internal function called by \code{bar_error} and \code{strip_error}.
}
