% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{panel.errbars1}
\alias{panel.errbars1}
\title{Internal function for displaying error bars in lattice plots.}
\usage{
panel.errbars1(x, y0, y1, ewidth = 0)
}
\arguments{
\item{x}{A numeric vector with the x positions for the bars.}

\item{y0}{A numeric vector with the lower limits of the bars.}

\item{y1}{A numeric vector with the upper limits of the bars.}

\item{ewidth}{An integer.}
}
\description{
\code{panel.errbars1} is an internal function called by \code{panel.errbars1}.
}
