% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{predict_inv}
\alias{predict_inv}
\title{Given y solve for x in a simple linear model.}
\usage{
predict_inv(model, y)
}
\arguments{
\item{model}{A simple linear model object (class lm).}

\item{y}{A numerical scalar, the value of the outcome for which we want to calculate the predictor x.}
}
\value{
The estimated value of the predictor.
}
\description{
\code{predict_inv} Calculates the value the predictor x that generates value y with a simple linear model.
}
