% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_functions.R
\name{pubh_hux}
\alias{pubh_hux}
\title{Displaying data frames and table of coefficients as huxtables}
\usage{
pubh_hux(data, label = "Parameter", arn = FALSE, digits = 2)
}
\arguments{
\item{data}{A data frame.}

\item{label}{A Character used to name the predictors in a table of coefficients.}

\item{arn}{Logical, should row names be added as a column \code{glm_coef}?}

\item{digits}{Number of digits to round numerical values (not including p-values).}
}
\value{
A huxtable.
}
\description{
\code{pubh_hux} extends \code{huxtable} objects by addin column names and rounding numeric cells.
}
\examples{
require(huxtable)
require(dplyr)
require(sjlabelled)

data(birthwt, package = "MASS")
birthwt <- birthwt \%>\%
  mutate(
    smoke = factor(smoke, labels = c("Non-smoker", "Smoker")),
    race = factor(race, labels = c("White", "African American", "Other"))
  ) \%>\%
  var_labels(
    bwt = 'Birth weight (g)',
    smoke = 'Smoking status',
    race = 'Race'
  )

model_norm <- lm(bwt ~ smoke + race, data = birthwt)

glm_coef(model_norm, labels = model_labels(model_norm)) \%>\%
  pubh_hux(arn = TRUE) \%>\%
  theme_article()
}
\seealso{
\code{\link[huxtable]{as_hux}}.
}
