% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{rank_leverage}
\alias{rank_leverage}
\title{Ranks observations by leverage.}
\usage{
rank_leverage(x)
}
\arguments{
\item{x}{A numeric variable. Missing values are removed by default.}
}
\value{
A data frame ranking observations by their leverage around the mean.
}
\description{
\code{rank_leverage} ranks observations by their leverage (influence) on the arithmetic mean.
}
\examples{
x <- rnorm(10, 170, 8)
x
mean(x)
rank_leverage(x)

x <- rnorm(100, 170, 8)
mean(x)
head(rank_leverage(x))
}
\seealso{
\code{\link{jack_knife}}.
}
