% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{rel_dis}
\alias{rel_dis}
\title{Relative Dispersion.}
\usage{
rel_dis(x)
}
\arguments{
\item{x}{A numerical variable. NA's observations are removed by default.}
}
\value{
The coefficient of variation (relative dispersion).
}
\description{
Calculates the coefficient of variation (relative dispersion) of a variable. The relative dispersion
is defined as the standard deviation over the arithmetic mean.
}
\examples{
height <- rnorm(100, 170, 8)
rel_dis(height)
}
\author{
Josie Athens, Department of Preventive and Social Medicine, University of Otago, New Zealand.
}
