% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_functions.R
\name{round_pval}
\alias{round_pval}
\title{Rounding p-values.}
\usage{
round_pval(pval)
}
\arguments{
\item{pval}{vector of p-values, numeric.}
}
\description{
\code{round_pval} is an internal function called by \code{glm_coef} to round p-values from model coefficients.
}
