% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{stats_quotes}
\alias{stats_quotes}
\title{Internal function to calculate descriptive statistics.}
\usage{
stats_quotes(x, data2, digits = 2)
}
\arguments{
\item{x}{a numeric variable}

\item{data2}{A data frame where \code{x} can be found.}

\item{digits}{Number of digits for rounding.}
}
\description{
\code{stats_quotes} is an internal function called by \code{estat}.
}
\author{
Josie Athens, Department of Preventive and Social Medicine, University of Otago, New Zealand.
}
