% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vassey}
\alias{Vassey}
\title{Case-control study of oral contraceptives and breast cancer.}
\format{A data frame with 863 rows and 2 variables:
\describe{
\item{OC}{Exposure to oral contraceptives, factor with levels "No" and "Yes".}
\item{Cancer}{Diagnosed with breast cancer, factor with levels "No" and "Yes".}
}}
\source{
Vessey M, Baron J, Doll R, Mc Pheson K, Yeates D (1983) Oral contraceptives and breast cancer:
Final report of an epidemiological study. Brit J Cancer, 47:455-462.
}
\usage{
Vassey
}
\description{
A data set about a case control study of oral contraceptives and breast cancer.
The cases were of recently diagnosed and histologically proven breast cancer in women aged 16-50 years,
in certain hospitals. The controls were married women inpatients in the same hospital,
who had acute medical or surgical conditions.
}
\examples{
contingency(Cancer ~ OC, data = Vassey, method = "case.control")
}
\keyword{datasets}
