% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_functions.R
\name{diag_test}
\alias{diag_test}
\title{Diagnostic tests from variables.}
\usage{
diag_test(object = NULL, formula = NULL, data = NULL, ...)
}
\arguments{
\item{object}{When chaining, this holds an object produced in the earlier portions of the chain. Most users can safely ignore this argument. See details and examples.}

\item{formula}{A formula with shape: outcome ~ predictor (see details).}

\item{data}{A data frame where the variables in the \code{formula} can be found.}

\item{...}{Further arguments passed to \link{epi.tests}.}
}
\description{
\code{diag_test} is a wrap function that calls \link{epi.tests} from package \code{epiR}.
It computes sensitivity, specificity and other statistics related with screening tests.
}
\details{
For the \code{formula}, the outcome is the gold standard and the explanatory variable is the new (screening) test. See examples.
}
\examples{
## We compare the use of lung’s X-rays on the screening of TB against the gold standard test.
Freq <- c(1739, 8, 51, 22)
BCG <- gl(2, 1, 4, labels=c("Negative", "Positive"))
Xray <- gl(2, 2, labels=c("Negative", "Positive"))
tb <- data.frame(Freq, BCG, Xray)
tb <- expand_df(tb)

tb \%>\%
  diag_test(BCG ~ Xray)
}
\seealso{
\code{epiR::epi.tests}
}
