% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kirkwood}
\alias{Kirkwood}
\title{Body weight and plasma volume.}
\format{A labelled data frame with 8 rows and 3 variables:
\describe{
\item{subject}{Subject ID.}
\item{weight}{Body weight in kg.}
\item{volume}{Plasma volume in litres.}
}}
\source{
Kirkwood, BR and Sterne, JAC (2003) Essential Medical Statistics. Second Edition. Blackwell.
}
\usage{
Kirkwood
}
\description{
Body weight and plasma volume in eight healthy men.
}
\examples{
data(Kirkwood)

Kirkwood \%>\%
  gf_point(volume ~ weight) \%>\%
  gf_lm(col = "indianred3", interval = "confidence", fill = "indianred3") \%>\%
  axis_labs()
}
\keyword{datasets}
